Add tests for feature layer
