+++ dcv/storage.py
@@ -1,9 +1,22 @@
+import inspect
 from urllib.parse import urljoin
 
 import requests
 
 import settings
-from dcv.utils import qualify_key
+
+
+def qualify_key(func):
+    def wrapper(*args, **kwargs):
+        key = args[0]
+        default_namespace = inspect.signature(func).parameters.get('namespace').default
+        namespace = kwargs.get('namespace', default_namespace)
+        qualified_key = f'{namespace}:{key}'
+        args = list(args)
+        args[0] = qualified_key
+        return func(*args, **kwargs)
+
+    return wrapper
 
 
 @qualify_key
