+++ tests/test_layer.py
@@ -1,0 +1,3 @@
+def test_shapefile_is_downloaded(feature_layer):
+    feature_layer.download_shapefile()
+    assert feature_layer.layer_file.exists()
